+++ src/main/java/lotto/model/ticket/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
