+++ src/test/java/lotto/model/error/InputValidatorTest.java
@@ -1,10 +1,11 @@
-package lotto.model;
+package lotto.model.error;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
 import lotto.error.InputError;
+import lotto.error.InputValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
